Tratamento de exceção

+++ jpa-spring/src/main/java/com/jpa_spring/jpa_spring/services/UserService.java
@@ -2,7 +2,12 @@ package com.jpa_spring.jpa_spring.services;
 
 import com.jpa_spring.jpa_spring.entities.User;
 import com.jpa_spring.jpa_spring.repositories.UserRepository;
+import com.jpa_spring.jpa_spring.services.exceptions.DataBaseException;
+import com.jpa_spring.jpa_spring.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,7 +25,7 @@ public class UserService {
 
     public User findById(Integer id){
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 
     }
 
@@ -29,13 +34,25 @@ public class UserService {
     }
 
     public void delete(Integer id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        } else {
+            try{
+                repository.deleteById(id);
+            } catch (DataIntegrityViolationException e){
+                throw new DataBaseException(e.getMessage());
+            }
+        }
     }
 
     public User update(Integer id, User user){
-        User entity = repository.getReferenceById(id);
-        updateData(entity, user);
-        return repository.save(entity);
+        try {
+            User entity = repository.getReferenceById(id);
+            updateData(entity, user);
+            return repository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User user) {
